Feature/ros noetic port (#151)
* Port to ROS Noetic/Ubuntu 20.04/Python 3

* Include a header to fix a lint issue

* Include a header to fix a lint issue

* noetic testing

* autoware_msgs discrepancy fix with noetic

* dependency fixes

* dns error fix in vm

* cleaned up comments

* test with development tags

* develop tags

* develop changes

* unit test changes

* roslint fixes

* more roslint changes

* roslint changes

* update to noetic/develop

* relase tags

* Update CMakeLists.txt files to use C++14

* Address code review feedback

* Make explicit selection of OpenGL library; put back C++ selection in cmake

+++ core_perception/trafficlight_recognizer/include/trafficlight_recognizer/rate.h
@@ -31,7 +31,7 @@ public:
   {
     assert(hz >= 1);
     float microsec = 1e6 / static_cast<float>(hz);
-    sleeptime = pt::microseconds(microsec);
+    sleeptime = pt::microseconds(static_cast<long>(microsec));
     lastUpdate = pt::microsec_clock::local_time();
   }
